# Auto batch dispatch calls