--- a/packages/data/src/utils/emitter.js
+++ b/packages/data/src/utils/emitter.js
@@ -7,11 +7,11 @@ export function createEmitter() {
 	let isPending = false;
 	const listeners = new Set();
 	const notifyListeners = () => {
+		isPending = false;
 		// We use Array.from to clone the listeners Set
 		// This ensures that we don't run a listener
 		// that was added as a response to another listener.
 		Array.from( listeners ).forEach( ( listener ) => listener() );
-		isPending = false;
 	};
 
 	return {