## What?

Call me noob, but I recently just learned about "microtasks" in JavaScript 🤷‍♂️ and I thought it was the ideal way to batch dispatch (store) calls instead of asking the user to explicitly call the `registry.batch` function.

## Why?

Performance, my hope is that we'll have a very small number of breakage and improvements to re-rendering performance.

## How?

Store listeners are only called on a microstask: After all sync code is expected but before any other event in the JS loop.
This is still very experimental, I wanted to see how the e2e test react to this change.