@Mamaduka 2022-03-17
This is interesting. I had no idea about "microtasks" as well 😅