@youknowriad 2022-03-17
I see a consistent and significant improvement in typing performance

trunk: 41.82 ms
this branch: 34.3 ms

I also see a loss in first block rendering performance

trunk: 6056.96 ms
this branch: 4783.52 ms

---
I think just from the typing performance improvement, it's obvious that this approach is worth considering and pushing further. There's a number of failing e2e tests that might rely on "sync" behavior. We need to check these.

For the "first block rendering", I'm not sure how to explain it yet. My guess is that the first block renders later now but that might not necessary mean a loss in performance. Otherwise, something that was happening after first block is now happening before. (But we need to dive deeper here)