@youknowriad 2022-03-17
Looks like after my fix in [3e1b9f3](https://github.com/WordPress/gutenberg/pull/39538/commits/3e1b9f3c83ec4b74d99b376e2d47a569098fffc1) Typing is still being improved but first block load is not causing a dramatic loss in performance 🎉