@youknowriad 2022-03-18
Unfortunately, this is not going to work in React 17. Here's a codesandbox that explains the issue https://codesandbox.io/s/funny-leftpad-1jkybt?file=/src/App.js:66-89

- Open the code sandbox
- Type something in the put
- Move the caret to the beginning of the input
- Type something else

Notice the focus jump to the end of the input. Event though, the state update does happen properly and the component only rerenders one, I guess the micro-tasks break the expected order of things for React internally, creating the focus jump. There's a possibility that React18 behaves differently here and in which case we could resurface this but there's no guarantee.

What does this mean for this PR? It means this PR is going to break any controlled input/textarea if the onChange uses the data module and not react state. It's a somewhat big breaking change so I'm going to close this for now.

If you have other ideas here, let me know. For now I'm closing this exploration.