@jsnajdr 2022-03-21
> Unfortunately, this is not going to work in React 17.

I think this would break any version of React ever, not only 17. On a controlled component, the `setState` must happen synchronously inside the `onChange` event handler, otherwise the caret position is lost. When done in a microtask, the value update is considered "external" and moves the caret. You can try this:
```js
function CountingInput() {
  const [ value, setValue ] = useState( '' );
  useEffect( () => {
    setInterval( () => setValue( val => val + 'x' ), 1000 );
  }, [] );
  return <input value={ value } />
}
```
Every time you move the caret in the input, the next timer tick will move it to the end as the next `x` is added.

> If you have other ideas here, let me know.

The batching would have to be much more targeted, as a default it's too blunt a tool.