@youknowriad 2022-03-21
The idea of micro tasks is that it's "almost sync" meaning, it's done asynchronously but before any other event in the JS loop. I think React 18 uses microtasks internally to auto batch setState calls, my hope is that it might make using micro tasks for our own optimizations possible but it's just a guess/hope, I didn't really check React 18 or its code base.