@jsnajdr 2022-03-22
I did a bit of experimentation and debugging (with DOM breakpoints) I found out that the microtask itself is not a problem. If I create a vanilla JS `input` with an `input` event listener and perform `el.value = value` in a microtask, the caret position is not lost.

The problem is solely inside React. With a delayed `setState` in an `onChange` handler:
```js
function onChange( event ) {
  const newValue = event.target.value;
  Promise.resolve().then( () => { // microtask
    setValue( newValue ); // setState
  } );
}
```
and when changing the input, say, from `a` to `ba`, the order of events is:
1. The browser changes the input value from `a` to `ba` and updates the screen. This is solely browser and DOM at work, no React involved yet. At this moment the `el.value` DOM attribute becomes `ba`.
2. A bit later, when handling the `setState` in the microtask, React sets `el.value` to `a`, the old value (!)
3. And then immediately sets it back to the desired `ba`.

It's the temporary value reset to `a` that causes the caret jump.

Why React does this? While handling an `input` event on a DOM element, it calls a function called `enqueueStateRestore`. Then, when updating the `<input>` component, while doing `setState` a bit later, it will call a `restoreControlledState` function on the element, which sets `value = 'a'` using the old value of the `value` prop. And then it finally updates the element once more with the new prop value, `ba`.

I don't know what exactly this enqueue/restore logic does, but now I'm sure that this caret jumping is not a DOM quirk, but a React implementation detail and if React 18 is going to batch events using microtasks, it's going to change this logic as well.