@jsnajdr 2022-03-22
When typing inside a paragraph block, I see three actions being dispatched:
- `SELECTION_CHANGE` (caret has moved)
- `UPDATE_BLOCK_ATTRIBUTES` (block `content` has changed)
- `EDIT_ENTITY_RECORD` (propagate the changes from block editor to entity store)

The first two are already batched together with `registry.batch`. The third one is dispatched inside a store listener which detects a "blocks have changed in block editor" event and dispatches and action right there in the listener.

Implementing batching with microtasks seems to me to only change the batching implementation, it doesn't add any new batching that was not already there. If the observed performance improvement in typing is real, not a random coincidence, then it's probably caused by changed order in how various rerenders and updates are done. Look at this flamechat excerpt (measured with this PR applied):

<img width="799" alt="Screenshot 2022-03-22 at 10 19 35" src="https://user-images.githubusercontent.com/664258/159448433-df712032-2a35-44ed-bc96-efad42fbf090.png">

The first part, `handleChange`, is a rerender of the `RichText` component itself, triggered by `forceRender` [here](https://github.com/WordPress/gutenberg/blob/trunk/packages/rich-text/src/component/index.js#L161-L168). The `registry.batch()` call, although it's done before `forceRender`, is postponed into a microtask. Before this PR, the order was reversed: do a data store dispatch and synchronously do all updates triggered in listeners, and only then do `forceRender`. This reordering is the only substantial change I see here.