@youknowriad 2022-03-22
@jsnajdr does this mean that potentially we can add a microtask to RichText directly to reproduce the same effect here? 

I think there's still value in auto-batching outside RichText though (we do a lot of sequential dispatches on load for instance...), and we don't need to think about calling `batch` ourselves. But that's for another time.